+++ src/Service/ReportService.php
@@ -7,11 +7,15 @@ declare(strict_types=1);
 namespace DVDoug\Behat\CodeCoverage\Service;
 
 use Composer\InstalledVersions;
+use Composer\Semver\VersionParser;
 use SebastianBergmann\CodeCoverage\CodeCoverage;
 use SebastianBergmann\CodeCoverage\Report\Clover;
 use SebastianBergmann\CodeCoverage\Report\Cobertura;
 use SebastianBergmann\CodeCoverage\Report\Crap4j;
+use SebastianBergmann\CodeCoverage\Report\Html\Colors;
+use SebastianBergmann\CodeCoverage\Report\Html\CustomCssFile;
 use SebastianBergmann\CodeCoverage\Report\Html\Facade as HtmlFacade;
+use SebastianBergmann\CodeCoverage\Report\Html\Thresholds;
 use SebastianBergmann\CodeCoverage\Report\PHP;
 use SebastianBergmann\CodeCoverage\Report\Text;
 use SebastianBergmann\CodeCoverage\Report\Xml\Facade as XmlFacade;
@@ -48,14 +52,42 @@ class ReportService
                     $report->process($coverage, $config['target'], $config['name']);
                     break;
                 case 'html':
-                    $report = new HtmlFacade(
-                        $config['lowUpperBound'],
-                        $config['highLowerBound'],
-                        sprintf(
-                            ' and <a href="https://behat.cc">Behat Code Coverage %s</a>',
-                            InstalledVersions::getPrettyVersion('dvdoug/behat-code-coverage')
-                        )
-                    );
+                    if (InstalledVersions::satisfies(new VersionParser(), 'phpunit/php-code-coverage', '^9.0')) {
+                        $report = new HtmlFacade(
+                            $config['lowUpperBound'],
+                            $config['highLowerBound'],
+                            sprintf(
+                                ' and <a href="https://behat.cc">Behat Code Coverage %s</a>',
+                                InstalledVersions::getPrettyVersion('dvdoug/behat-code-coverage')
+                            )
+                        );
+                    } else {
+                        $thresholds = Thresholds::from(
+                            $config['lowUpperBound'],
+                            $config['highLowerBound'],
+                        );
+                        $colors = Colors::from(
+                            $config['colors']['successLow'],
+                            $config['colors']['successMedium'],
+                            $config['colors']['successHigh'],
+                            $config['colors']['warning'],
+                            $config['colors']['danger'],
+                        );
+                        if ($config['customCSSFile']) {
+                            $customCss = CustomCssFile::from($config['customCSSFile']);
+                        } else {
+                            $customCss = CustomCssFile::default();
+                        }
+                        $report = new HtmlFacade(
+                            sprintf(
+                                ' and <a href="https://behat.cc">Behat Code Coverage %s</a>',
+                                InstalledVersions::getPrettyVersion('dvdoug/behat-code-coverage')
+                            ),
+                            $colors,
+                            $thresholds,
+                            $customCss
+                        );
+                    }
                     $report->process($coverage, $config['target']);
                     break;
                 case 'php':